requirements.txt
annotated-types==0.5.0
anyio==3.7.1
app==0.0.1
catboost==1.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.3
fastapi==0.101.1
filelock==3.12.2
fonttools==4.42.0
fsspec==2023.6.0
graphviz==0.20.1
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
joLib==0.0.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pipeline==0.1.0
plotly==5.16.1
pydantic==2.1.1
pydantic_core==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
sklearn==0.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
xgboost==1.7.6
zipp==3.16.2
